backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
pydantic-settings
python-jose[cryptography]
passlib[argon2]
argon2-cffi
alembic
Pillow
pandas
openpyxl
fastapi-mail
